Refactored to treat SizeBuckets as ABC
